Remove incompatible npm script
The "clean" script use "rm" command which is not available on windows
for example.

A "rimraf" node module exists to solve this compatibility issue, but
in this use case where the command was used to remove the node_modules
folder it fails because the rimraf module has to delete itself.